Move StartPathfinding into Vertex as well as CalculateVertexParameters and DeterminePath into EndPoint
