--- Assets/Scripts/Pathfinding.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: e9e77034cffb4da98ff4c705285a46a9
-timeCreated: 1580488941
